Move GetAffinityTerms functions from pkg/scheduler/framework to staging repo
