+++ staging/src/k8s.io/kube-scheduler/framework/types_test.go
@@ -16,7 +16,15 @@ limitations under the License.
 
 package framework
 
-import "testing"
+import (
+	"testing"
+
+	"github.com/google/go-cmp/cmp"
+
+	v1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/util/sets"
+)
 
 type hostPortInfoParam struct {
 	protocol, ip string
@@ -227,6 +235,48 @@ func TestHostPortInfo_Check(t *testing.T) {
 			}
 			if hp.CheckConflict(test.check.ip, test.check.protocol, test.check.port) != test.expect {
 				t.Errorf("expected %t; got %t", test.expect, !test.expect)
+			}
+		})
+	}
+}
+
+func TestGetNamespacesFromPodAffinityTerm(t *testing.T) {
+	tests := []struct {
+		name string
+		term *v1.PodAffinityTerm
+		want sets.Set[string]
+	}{
+		{
+			name: "podAffinityTerm_namespace_empty",
+			term: &v1.PodAffinityTerm{},
+			want: sets.Set[string]{metav1.NamespaceDefault: sets.Empty{}},
+		},
+		{
+			name: "podAffinityTerm_namespace_not_empty",
+			term: &v1.PodAffinityTerm{
+				Namespaces: []string{metav1.NamespacePublic, metav1.NamespaceSystem},
+			},
+			want: sets.New(metav1.NamespacePublic, metav1.NamespaceSystem),
+		},
+		{
+			name: "podAffinityTerm_namespace_selector_not_nil",
+			term: &v1.PodAffinityTerm{
+				NamespaceSelector: &metav1.LabelSelector{},
+			},
+			want: sets.Set[string]{},
+		},
+	}
+
+	for _, test := range tests {
+		t.Run(test.name, func(t *testing.T) {
+			got := getNamespacesFromPodAffinityTerm(&v1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "topologies_pod",
+					Namespace: metav1.NamespaceDefault,
+				},
+			}, test.term)
+			if diff := cmp.Diff(test.want, got); diff != "" {
+				t.Errorf("Unexpected namespaces (-want, +got):\n%s", diff)
 			}
 		})
 	}
